Fix temp directory path in unittests
#7060 introduced Bandit and made some changes to satisfy its checks. One of those changes was to use `tempfile.gettempdir()` rather than a hard-coded `/tmp` in Ask CFPB exports.

This had the effect of breaking our unittests when the `TMPDIR` envvar isn’t set to `/tmp` itself, because we hard-coded that path in the unit tests, as [@chosak noted](https://github.com/cfpb/consumerfinance.gov/pull/7060#issuecomment-1130320885).

This change fixes that by reusing the value of `tempfile.gettempdir()` for that test.