Add simple chart projected data label
We've been relying on a "plot line" label which draws a vertical plot
line and labels it with some text. Unfortunately, if a user zooms to an
area of the chart that doesn't contain the plot line, the label is also
removed. We want the projected data label to always be visible so we
instead render it as text positioned over the chart.