Convert email signups to snippets
This change converts our email signups to snippets, so that we can share the same email signup block data across multiple pages without so much duplication.

This includes a data migration that will either create a new `EmailSignUp` snippet object when it encounters an old `EmailSignUp` block on a page. This will likely result in a lot of email signups being created that are minor duplicates (slight differences in wording).

Internally, we have a list of the email signups we *should* have. This is likely the first step toward making that list reflect reality in the CMS. There will likely be some consolidation of `EmailSignUp` snippet objects that we could do programmatically.

This change also adds a common function to migrate email signups, this function can be removed in a subsequent change after this one has been deployed.