Remove pip package resolution dependency
This change removes our check for whether a wheel file is supported. This primarily targetted the use of Python 2 and 3, but now that we’re well past the transition to Python 3, this can go away.

This will remove our dependency on pip internals as well, decreasing the likelihood of future breakage (and making way to new things to break).

This change also removes our support for multiple Python versions in the zipfile creation.