Deprecate page archiving functionality (#7081)
This commit deprecates the CFGOVPage archiving functionality that was
implemented in 2020 but never launched. This removes the common
`is_archived` and `archived_at` fields on all pages, as well as the
related Django management commands `archive_pages` and
`restore_archive_pages`. It also removes all related frontend
components related to the display of archived pages.

See internal platform#4416 for some discussion on this change.