Remove utf8 param from search base URL
This changes removes the `utf8=✓` parameter from our search base URL. This will remove that parameter from our no-JS fallback link used by IE that links directly to the search.gov search page for CFPB. Its removal shouldn't impact JS-enabled search as there is a hidden input that adds the parameter on lines 55-57. 

In IE, with this parameter, we're seeing 400 Bad Request errors on following the link that appear down to the UTF-8 `✓` character. This is an attempt to solve that issue.