FIG: Custom content options, including info-unit-group (#7114)
* Use custom streamfield for FIG content

Previously, the FIG Content Page blocks (section, subsection, and level
3 subsection) used FullWidthText as their content block. Instead, we'll
use a StreamBlock with a custom selection of content blocks. For now,
I've taken an educated guess about what kinds of content are likely to
be needed in this kind of page. If we decide to modify the list of
available blocks in the future, it won't require a migration in four
separate apps, the way it would if we continued to use FullWidthText.

* Add FIG content blocks to FIG section template

rather than relying on full_width_text to display the blocks

* Move FIG-section template into filing_instruction_guide app

* Remove duplicated template code