Mark page class abstraction test no-cover
This change marks the `page_cls._meta.abstract` test for no-coverage checking. The rationale is that I have not found an example page class in our codebase or in Wagtail’s upstream that is abstract that would meet this test. Rather than creating one for the purpose of covering a `continue` line, I am opting not to coverage-check this line.