Use the Blog to test email signups
This change to our functional tests will use the blog’s email signup flow rather than `/consumer-tools`.

Recently we modified email signups to support either a signup link directly to GovDelivery or our existing signup box on the page. `/consumer-tools` was one of the URLs switched to a signup link, thus it is no longer useful for testing the signup flow.

The test database is updated to add an email signup to `/about-us/blog` for testing.