FIG: Don't include a trailing dot in section IDs (#7151)
Previously, a section ID for a part of a FIG document would be somethign
like 3.1.2. (including the trailing dot). Now, the section ID for the
same section would be 3.1.2 (without a trailing dot).

With this PR, an anchor link to an individual section of the FIG won't
include a trailing dot (which was weird), but the body of the page will
still have the dot in headers and the table of contents. Now, the dot
is added by the Jinja template.