Add a `wait` to email signup functional test
This small change adds a `wait` to our email signup tests between the signup submission and checking for the success notification. Previously, we were seeing occasional failures to find the success notification, and observing the results in Cypress, it seems to be because Cypress gets to the success check before the success notification has rendered.

This wait seems to fix it against staging, which is the only place I was able to reproduce the issue with any regularity.