Add FIG version status and effective dates
Filing Instruction Guide pages have statuses and effective dates that
are propagated to the front end via a banner at the top of the page
and a pill in the left-hand nav.

The banner and pill are only shown if the page status is anything other
than "current". The pill uses our "tag" pattern and modifier classes
have been added to `tag.less`.

https://github.local/CFPB/1071-data-collection/issues/178
https://github.local/CFPB/1071-data-collection/issues/185