Fix `data-option` lookup for categories
This change fixes our `data-option` lookup from the human-friendly string for categories that are don’t quite line up as kebab-cased versions of the string.

In this case, “Data, Research, and Policy” gets converted to `data,-research,-and-policy`, when the actual `data-option` value is `data-research-policy`. A similar problem occurs with “Director’s Notebook”.

This fix provides a helper function that will use a regular expression to replace dropped words (“and”) and characters (“,”, “’”) before kebab-casing the resulting string.

I’ve made use of this function in all the places where we lookup categories and topics from human-friendly strings.