Remove direct RPM install from NodeSource
It looks like there’s an upstream issue with `nodejs-16.13.1-1nodesource.x86_64.rpm` at NodeSource. Installing this RPM directly results in a “Package does not match intended download” error from rpm.

It’s [not just us having the issue](https://github.com/nodesource/distributions/issues/1432), and from the directory listing, that file was updated yesterday.

Removing the direct install, which was added in #6747 mitigates the issue, and I don’t see a return of that error message without it that #6747 intended to fix.