Remove additional redundant title check
This change removes some additional checks for the `title` parameter in the URL after the `title` field is checked in the filterable list controls. The `title` parameter check was failing occasionally due to URL-encoding of the title. The title parameter check seems redundant given that we’re checking the title field of the control after applying the filters.