Add wagtail-editable-help to edit help text
This commit adds the wagtail-editable-help [0] package as a way to
make Wagtail help text editable. This initial commit only adds the
package and manually changes two pieces of text to be editable.

Ideally there would be a way to automatically add all help text instead
of having to do this one at a time; this could maybe be done via
script.

[0] https://github.com/wagtail/wagtail-editable-help