Automatically import FIG data points from JSON
Upon save, preview, or publish, a FIGContentPage now downloads JSON
(following a specific structure representing FIG data points) from a
specified URL, parses, and saves it to the page. The content editor
places a FigDataPointsBlock in whatever Section of the page they want
the data points to appear in.

On page load, the data point info is placed into the data points
template in the section specified by the content editor. Links to the
data points are added to the table of contents.

To do: Find a nicer way to validate the JSON schema. Consider
using Wagtail Snippets to hold the JSON data, rather than downloading
from a separate location.