Support non-legacy content on LegacyBlogPages (#7284)
Currently LegacyBlogPages (and LegacyNewsroomPages) can only contain
content blocks that are either RawHTMLBlocks or ReusableTexts.
Non-legacy BlogPages (and NewsroomPages) support a much broader set of
content options.

As part of an effort to migrate LBPs and LNPs, we want to migrate the
content on those pages to non-legacy layouts. This commit adds support
for non-legacy content to those page types, which will give us the
ability to migrate the data in-place on existing pages. (A
subsequent step will involve actually changing the type of LBP/LNPs
to BP/NPs, but first we can/should migrate the data they contain.)