Make Policy & Compliance blog choice value consistent
This changes the `policy_compliance` ref value to `policy-compliance`, for consistency with the other values. This will enable more consistent programmatic selection of categories in our filterable list tests.

This change includes a data migration that will migrate all existing `policy_compliance` categorizations to `policy-compliance`.

This requires a re-index of the search index after the migration is performed.