Use a dedicated Node image for the frontend build
This change splits up the `cfgov-frontend-builder` image in our `Dockerfile` so that we use a dedicated `node:16-alpine` image for the steps that require Node.js instead of installing the `nodejs` package into the `python:3.8-alpine` image. This should let specify separately the version of Node that’s used in the build.