Ensure FIG data field tables have correct number of columns
Every row in the table needs to have the same number of columns so we
iterate over the header row and print the value of the current row at the
given index. If the row has no value at that index, print an empty string.

See https://github.local/CFPB/1071-data-collection/issues/302