Cleanup filterable form language list
As written, the filterable form language option is populated from an
aggregation added to the Elasticsearch query. The code currently
supports the case when the aggregation doesn't exist in the results,
to handle a time period in the past (July 2021) where production
Elasticsearch indexes might exist without it. As enough time has
passed since then, and all search caches should have been rebuilt,
this logic (added in [0]) is no longer needed.

[0] https://github.com/cfpb/consumerfinance.gov/commit/e7269685a5833893ec4000d4a2bcc7a5ed16c118