Configure default Python formatter for VSCode
PR #7325 configured the VSCode settings for this repository so that
Prettier is the default code formatter. This doesn't work for Python
files. This commit changes the VSCode settings to set the Python
extension as the default formatter; this lets VSCode users use
black to format Python files.

This commit also adds documentation about these VSCode extensions.