Improve 1071 FIG table of contents scrollspy accuracy
The table of contents in the sidenav highlights the section that the
user is currently viewing. It was doing so by tracking the position of
each section heading in the main content. Now that some sections are
looooooooong, headings can be few and far between, causing the nav
to not get highlighted until a heading is reached.

The section container divs are now being tracked instead of just the
headings to make the scrollspy more accurate.

See https://github.local/CFPB/1071-data-collection/issues/291