Deprecate VSCode configuration
As written, the current Visual Studio Code configuration file
establishes Prettier as the default code formatter, and automatically
formats files on save. This causes problems if you edit HTML files in
VSCode, because these files get reformatted in undesirable ways.

This config file was just added last month in PR 7325, and then had
to be further customized in PR 7343 to prevent Prettier from editing
Python files. I propose that we remove editor configuration from this
repository; we can document standards and enforce them with precommit
and/or GitHub Actions.