Fix: Item introduction categories don't link (#7360)
Item introduction categories originally linked back to a filtered
version of their parent filter, when the item introduction configuration
option "Show category" is checked.

This was broken by this commit [0] back in 2018 (by yours truly).

We use "Show category" sparingly on item introductions. An example
production page with this option checked is this one:

https://www.consumerfinance.gov/foia-requests/foia-electronic-reading-room/2017-student-banking-report-correspondence/

Note that currently the category renders as a plain text "Record" that
doesn't link anywhere. With this commit, the category renders as a bold
link that links back to the parent filterable list.

[0] https://github.com/cfpb/consumerfinance.gov/commit/705d8f0f9cad6f4aa807bd2f9147d4649fcc96aa