Change content-dependent filterable language tests
Our single and multiple language filterable blog posts tests were looking for specific text from specific blog posts, which breaks if new blog posts are published and the ones the test is looking for disappear behind pagination.

This changes the tests to look for post counts based on the `lang` attribute of the post preview in the filter results.