Bump some Wagtail packages to support 3.x
This commit bumps several of our Wagtail-related packages to newer
versions that support Wagtail 3.x while still supporting our current
version of Wagtail, 2.15.5:

wagtail-autocomplete from 0.8.0 to 0.9.0
wagtail-flags from 5.1.0 to 5.2.0
wagtail-inventory from 1.5 to 1.6
wagtail-sharing from 2.6 to 2.7

Of these 4, we maintain all but wagtail-autocomplete, and can vouch
for their continued functionality. Unfortunately we don't have a
functional test that checks wagtail-autocomplete. To verify it
manually, you can edit an Ask CFPB page [0] and scroll down to the
"related questions" field, and verify that it still works properly.

After this change, we still have one Wagtail-related package that
doesn't support Wagtail 3.x, which is our wagtail-treemodeladmin [1].
That support needs to be added and its version bumped in a future PR.

[0] http://localhost:8000/admin/pages/8904/edit/
[1] https://github.com/cfpb/wagtail-treemodeladmin/issues/33