Scroll to blocks in Cypress tests
In Cypress tests we check for block visibility using

.should('be.visible')

but this assumes that the block is visible within the browser screen
dimensions. In Wagtail 3 the admin UI changes significantly, and the
spacing changes mean that some blocks may be off screen.

This commit adds

.scrollIntoView()

before checking whether a block is visible.