Upgrade to Wagtail 3.0.3 (#7423)
* Bump Wagtail version from 2.16.3 to 3.0.3

* Update Wagtail module paths for 3.x

This commit updates the imports of Wagtail modules to be compatible
with the new organization in Wagtail 3.x. These changes are the result
of running Wagtail's updatemodulepaths command, as documented:

https://docs.wagtail.org/en/latest/releases/3.0.html#changes-to-module-paths

* Use FieldPanel instead of StreamFieldPanel

StreamFieldPanel has been deprecated in 3.0:

https://docs.wagtail.org/en/latest/releases/3.0.html#removal-of-special-purpose-field-panel-types

* Use FieldPanel instead of ImageChooserPanel

ImageChooserPanel has been deprecated in 3.0:

https://docs.wagtail.org/en/latest/releases/3.0.html#removal-of-special-purpose-field-panel-types

* Use FieldPanel instead of DocumentChooserPanel

DocumentChooserPanel has been deprecated in 3.0:

https://docs.wagtail.org/en/latest/releases/3.0.html#removal-of-special-purpose-field-panel-types

* Use FieldPanel instead of SnippetChooserPanel

 SnippetChooserPanel has been deprecated in 3.0:

https://docs.wagtail.org/en/latest/releases/3.0.html#removal-of-special-purpose-field-panel-types

* Use FieldPanel instead of PageChooserPanel

PageChooserPanel has been (mostly) deprecated in 3.0:

https://docs.wagtail.org/en/latest/releases/3.0.html#removal-of-special-purpose-field-panel-types

* Fix FieldPanel usage for Wagtail 3.x

In Wagtail 2.x, the second positional argument to FieldPanel was its
heading, but starting in 3.x you need to specify it as a keyword
argument (`heading="whatever"`).

* Update indexed field on IndexedPageRevision

In Wagtail 3.0, PageRevision models have a content field, not a
content_json field.

* Workaround for fbjs global undefined issue (#7446)

* Scroll to blocks in Cypress tests

In Cypress tests we check for block visibility using

.should('be.visible')

but this assumes that the block is visible within the browser screen
dimensions. In Wagtail 3 the admin UI changes significantly, and the
spacing changes mean that some blocks may be off screen.

This commit adds

.scrollIntoView()

before checking whether a block is visible.

* Update migration utils for content JSONField

In Wagtail 3.0+, the PageRevision content_json TextField has been
changed to a content JSONField. This commit updates our custom
migration utils to support that change.

* Update uses of PageRevision content field

In Wagtail 3.0+, the PageRevision content_json TextField has been
changed to a content JSONField. This commit updates the places in our
code where we referenced that field.

* Avoid unsafe chaining of Cypress functions

https://docs.cypress.io/api/commands/scrollintoview

Co-authored-by: Ans <anselmbradford@users.noreply.github.com>