Deprecate Chinese (Simplified) language
This commit deprecates the use of "Chinese (Simplified)" as a
support language for Wagtail pages. All existing pages marked as that
language get changed to "Chinese (Traditional)" instead. In practice,
there are no live pages like this right now, so this only applies to
previously deleted pages in the Trash.

The language code being deprecated is zh-Hans; pages are being changed
to code zh-Hant.

This addresses internal D&CT#141.