Fix retirement Spanish about page translations
Because of [a known issue with translation-marker-discovery in our Jinja templates with our front-end functional call style](https://cfpb.github.io/consumerfinance.gov/translation/#do-not-include-spaces-between-the-parentheses-and-the-string-in-jinja2-templates), the existing Spanish translations on this page have been broken.

The issue is, if a `_("foo")` call in a template is instead written as `_( "foo" )` with spaces around the `_()` argument, Django's message scanning with `makemessages` will miss the string.