Render translation links in their own language
This commit introduces a new Jinja {% language %} tag modeled
after the Django {% language %} tag, and uses it to ensure that
page translation links always get rendered in the destination language
(e.g. links to a Spanish translation always say "Español" regardless
of the language of the page being viewed).