Move login-related models to the `login` app
This change moves our three login-related models, `PasswordHistoryItem`, `FailedLoginAttempt`, and `TemporaryLockout` out of `v1` and into our separate `login` app.

To move these models, I’ve decided to simply copy the data over from the `v1` models rather than attempt to use more complex `migrations. AlterModelTable`-based migrations or to try to map the `login` models back onto `v1` tables. I chose this for future clarity — so that what happened when the models move is clear, what the present state of the models are is clear, etc.

The migrations are three steps:

1. `v1.0225_prepare_login_models_to_move` changes the related name to `User` from the default on all three models to prefix it with `v1`.
2. `login.0001_add_login_models` adds the models with a default related name to the `login` app.
3. `login.0002_populate_login_models` creates instances of the `login` models for each instance of the `v1` model and deletes the `v1` model.
4. `v1.0226_remove_login_models` removes the models from `v1`.

Hopefully this is straight-forward in what it is doing.