Define WAGTAILADMIN_BASE_URL setting
As of Wagtail 3.0, we need to define a Django setting named
WAGTAILADMIN_BASE_URL [0] to tell Wagtail the base URL of the
running website (example: https://www.consumerfinance.gov).

Without defining this, Django reports this warning when it runs:

?: (wagtailadmin.W003) The WAGTAILADMIN_BASE_URL setting is not defined
HINT: This should be the base URL used to access the Wagtail admin site.
Without this, URLs in notification emails will not display correctly.

This commit sets a default of "http://localhost:8000" for this value,
or uses the environment variable WAGTAILADMIN_BASE_URL if it is defined.

Ideally we would be able to put "http://localhost:8000" in our
settings.local and rely on `os.environ["WAGTAILADMIN_BASE_URL"]` in our
settings.production. Due to uncertainty around use of cf.gov
production settings in Docker-based environments, though, we can't
guarantee that environment variable will be set.

[0] https://docs.wagtail.org/en/stable/reference/settings.html#wagtailadmin-base-url