Translation links style iteration
This commit makes two minor style changes to how we render Wagtail
page translation links:

1. Font weight is set to 500.
2. The links <ul> is wrapped in a <div class="block block__sub"> to
ensure 30px of collapsible margin above and below.

These changes come from internal D&CP#122#issuecomment-328946.

Note that this change does not address the placement of the links on
the page; that is to come in a subsequent PR.

To test locally, `yarn run build` then visit:

- template debug view: http://localhost:8000/admin/template_debug/v1/translation_links/
- BlogPage: http://localhost:8000/about-us/blog/know-your-rights-and-protections-when-it-comes-to-medical-bills-and-collections/
- BrowsePage: http://localhost:8000/coronavirus/mortgage-and-housing-assistance/help-for-homeowners/
- LearnPage: http://localhost:8000/coronavirus/mortgage-and-housing-assistance/renter-protections/
- NewsroomPage: http://localhost:8000/about-us/newsroom/cfpb-and-ny-attorney-general-sue-repeat-offender-moneygram-for-leaving-families-high-and-dry/
- SublandingPage: http://localhost:8000/coronavirus/mortgage-and-housing-assistance/