Add an active users report
As a way to understand who we need to inform about upgrades (as well as maybe helping when we want to explore more about our users), I’ve added a new Wagtail report that just lists (in Excel-exportable form) the username, first, last, and email of each active user (has logged in within the last 90 days).

This puts in the Wagtail admin the information we’d usually go to the Django admin for, but Excel-exportable.