Add missing secondary nav JS to Spanish pages (#7526)
Currently any pages that inherit from SecondaryNavigationJSMixin only
include the secondary navigation JS on English language pages. But
the secondary navigation is also used for the "In this section"
expandable component rendered on BrowsePages on mobile.

So, on these Spanish pages the JS is currently missing, causing the
expandable to render as open (and be un-closeable). For example,
visit this page at a mobile width and notice the fixed "En esta
sección" component:

https://www.consumerfinance.gov/es/herramientas-del-consumidor/prestamos-del-dia-de-pago/respuestas/

Compare to a comparable English page:

https://www.consumerfinance.gov/consumer-tools/reverse-mortgages/answers/