Consolidate secondary navigation JS
Currently we add the secondary navigation JS to certain page types via
Python code, but this could be added once as part of the common
layout-side-nav.html template.

This change simplifies a few templates so that we can add it once.

A few pages to compare against to verify that this doesn't change any
existing behavior:

https://www.consumerfinance.gov/consumer-tools/debt-collection/answers/key-terms/
https://www.consumerfinance.gov/rules-policy/regulations/1002/
https://www.consumerfinance.gov/data-research/cfpb-research-conference/2022-cfpb-research-conference/