Deprecate ContactExpandables
This commit deprecates the ContactExpandable StreamField block as well
as its related ContactExpandableGroup block. Both of these blocks were
only available on LearnPages and were implemented to support an old
design of the Contact Us page which is no longer active.

See 2ad4b3c7f1e606ea7cf8437e2076ec8cc126a467 for the commit where
most of this code was added, as well as internal el-camino#21 for
context of why it was added at the time.

There are currently no pages on the website that use any of the
ContactExpandable code, so it is being removed in the interest of
simplification.

Additionally, this change also simplifies the Expandable module by
removing several options which are also unused: ContactAddress,
ContactEmail, and ContactPhone. The Expandable module has supported
these options since its creation in 2015
(9f696164473ba98a61eb1c7009326f50ab2d0f45), but there are no live pages
on the site using them.